training for duration model

+++ clustering/clusters.py
@@ -1,4 +1,5 @@
 import logging
+import sys
 import joblib
 import argparse
 from pathlib import Path
@@ -7,11 +8,12 @@ from sklearn.cluster import KMeans
 import torch
 from tqdm import tqdm
 
-
-BNS_PATH = Path(__file__).parent / "bns"
-
-
-import torch
+logging.basicConfig(
+    level=logging.INFO,
+    format="[%(filename)s] %(levelname)s: %(message)s",
+    handlers=[logging.StreamHandler(sys.stdout)],
+    force=True,  # important if something else configured logging first (common in notebooks/uv/etc.)
+)
 
 
 @torch.no_grad()
@@ -61,13 +63,13 @@ def run_lengths(inv: torch.Tensor):
     return values, lengths
 
 
-def load_bns_and_clusters():
+def load_bns_and_clusters(bns_path):
 
     chunks = []
     Ts = []
     file_ids = []
 
-    for f in BNS_PATH.glob("*.pt"):
+    for f in bns_path.glob("*.pt"):
         bn = torch.load(f).transpose(0, 1)
         T, _ = bn.shape
         Ts.append(T)
@@ -134,12 +136,19 @@ def spectral_order(centroids: torch.Tensor, k_nn: int = 8, sigma: float | None =
 if __name__ == "__main__":
 
     argument_parser = argparse.ArgumentParser()
+    argument_parser.add_argument("--bns_path", default="./data/LJSpeech-1.1/latents")
+    argument_parser.add_argument("--out_path", default="./data/LJSpeech-1.1")
     argument_parser.add_argument(
-        "--bns_path", default="../prosody_flow/data/LJSpeech-1.1/latents"
+        "--kmeans_object",
+        default="kmeans.joblib",
+        help="The name of the saved kmeans object. Should be in out_path",
     )
     argument_parser.add_argument(
-        "--out_path", default="../prosody_flow/data/LJSpeech-1.1"
+        "--embeddings_filename",
+        default="embeddings.pt",
+        help="The name of the saved embeddings object. Should be in out_path",
     )
+
     args = argument_parser.parse_args()
 
     chunks = []
@@ -147,34 +156,49 @@ if __name__ == "__main__":
     file_ids = []
 
     for f in tqdm(
-        Path(args.bn_path).glob("*.pt"),
-        desc=f"Fetching bns from {str(Path(args.bn_path))}...",
+        Path(args.bns_path).glob("*.pt"),
+        desc=f"Fetching bns from {str(Path(args.bns_path))}...",
     ):
-        bn = torch.load(f).transpose(0, 1)
+        bn = torch.load(f, map_location="mps")
+        if bn.ndim == 3:
+            bn = bn.squeeze(0)
+        bn = bn.transpose(0, 1)
         T, _ = bn.shape
         Ts.append(T)
         chunks.append(bn)
         file_ids.append(f.stem)
 
-    x = torch.cat(chunks)
-    labels, centroids, kmeans = cluster(x)
-
     out_path = Path(args.out_path)
-    logging.info(f"Saving embedding matrix to {str(out_path / 'embeddings.pt')}...")
-    torch.save(centroids, out_path / "embeddings.pt")
+    
+    Ts = torch.tensor([0] + Ts)
+    Ts = torch.cumsum(Ts, dim=-1)
 
-    Ts = np.array([0] + Ts)
-    Ts = np.cumsum(Ts)
+    torch.save(Ts.clone().contiguous(), out_path / "T.pt")
+
+    kmeans_path = out_path / args.kmeans_object
+    if not kmeans_path.exists():
+        x = torch.cat(chunks)
+        logging.info(f"Computing kmeans from bns...")
+        labels, centroids, kmeans = cluster(x)
+
+        logging.info(f"Saving kmeans model to {str(out_path / 'kmeans.joblib')}...")
+        joblib.dump(kmeans, out_path / "kmeans.joblib")
+
+        logging.info(f"Saving embedding matrix to {str(out_path / 'embeddings.pt')}...")
+        torch.save(centroids, out_path / "embeddings.pt")
+    else:
+        logging.info(f"Loading kmeans model from {str(kmeans_path)}...")
+        kmeans = joblib.load(str(kmeans_path))
+        labels = torch.from_numpy(kmeans.labels_)
+
+    
 
     emb_out_path = out_path / "emb_ids"
+    emb_out_path.mkdir(parents=True, exist_ok=True)
 
-    for start, end, file_id in enumerate(
-        tqdm(
-            zip(Ts[:-1], Ts[1:], file_ids),
-            desc=f"Saving embeddings to {str(emb_out_path)}",
-        )
+    for start, end, file_id in tqdm(
+        zip(Ts[:-1], Ts[1:], file_ids),
+        desc=f"Saving embeddings to {str(emb_out_path)}",
     ):
+        utt_labels = labels[start:end].clone().cpu()
         torch.save(labels[start:end], emb_out_path / f"{file_id}.pt")
-
-    logging.info(f"Saving kmeans model to {str(out_path / 'kmeans.joblib')}...")
-    joblib.dump(kmeans, out_path / "kmeans.joblib")
